refactor(svelte): move legend/ into a feature directory
Co-locate legend focus, filter, surface, and component modules under
legend/ with short feature-local names. Rewrite importers and mirror
tests under tests/legend/. Public legend-filter export names unchanged.
