+++ packages/svelte/src/lib/runtime/runtime.svelte.ts
@@ -15,7 +15,7 @@ import {
 import type { PortableSpec } from "@ggsvelte/spec";
 import { untrack } from "svelte";
 
-import type { LegendFilterClause } from "../legend-filter.js";
+import type { LegendFilterClause } from "../legend/filter.js";
 import type { ReadonlyZoomDomains } from "../interaction/interaction.js";
 import { isContainerWidthProp, resolvePlotSize } from "../assembly/layout.js";
 import { createPaintLedger, isPlotReady } from "./paint.js";
