+++ packages/svelte/tests/legend/renderer-mask.test.ts
@@ -2,8 +2,8 @@ import { describe, expect, it } from "vitest";
 
 import type { BatchInteractionMask, GeometryBatch, ThemeTokens } from "@ggsvelte/core";
 
-import BatchFocusHarness from "./fixtures/BatchFocusHarness.svelte";
-import { render } from "./helpers/render.js";
+import BatchFocusHarness from "../fixtures/BatchFocusHarness.svelte";
+import { render } from "../helpers/render.js";
 
 const theme = {
   ink: "#111111",
